Refactor pages for organization purposes
