+++ frontend/app/(authenticated)/question-repo/data-table-pagination.tsx
@@ -41,7 +41,7 @@ export function DataTablePagination<TData>({
                             <SelectValue placeholder={table.getState().pagination.pageSize} />
                         </SelectTrigger>
                         <SelectContent side="top" className="font-sans">
-                            {[10, 20, 30, 40, 50].map((pageSize) => (
+                            {[5, 10, 20, 30, 40, 50].map((pageSize) => (
                                 <SelectItem key={pageSize} value={`${pageSize}`}>
                                     {pageSize}
                                 </SelectItem>
